+++ java/src/hype/core/drawable/HDrawable.java
@@ -1351,8 +1351,7 @@ public abstract class HDrawable extends HNode<HDrawable>
 	 * @return This drawable.
 	 */
 	public HDrawable rotation(float deg) {
-		_rotationRad = deg * HConstants.D2R;
-		return this;
+		return rotationRad(deg * HConstants.D2R);
 	}
 	
 	/**
@@ -1361,18 +1360,28 @@ public abstract class HDrawable extends HNode<HDrawable>
 	 * @return This drawable's rotation in degrees.
 	 */
 	public float rotation() {
-		return _rotationRad * HConstants.R2D;
+		return rotatesChildren()? _firstChild.rotation() :
+			_rotationRad*HConstants.R2D;
 	}
 	
 	@Override
 	public HDrawable rotationRad(float rad) {
-		_rotationRad = rad;
+		if(rotatesChildren()) {
+			HDrawable d = _firstChild;
+			while(d != null) {
+				d.rotationRad(rad);
+				d = d.next();
+			}
+		} else {
+			_rotationRad = rad;
+		}
 		return this;
 	}
 	
 	@Override
 	public float rotationRad() {
-		return _rotationRad;
+		return (rotatesChildren() && _firstChild!=null)?
+			_firstChild.rotationRad() : _rotationRad;
 	}
 	
 	/**
@@ -1383,8 +1392,7 @@ public abstract class HDrawable extends HNode<HDrawable>
 	 * @return This drawable.
 	 */
 	public HDrawable rotate(float deg) {
-		_rotationRad += deg * HConstants.D2R;
-		return this;
+		return rotateRad(deg * HConstants.D2R);
 	}
 	
 	/**
@@ -1395,11 +1403,17 @@ public abstract class HDrawable extends HNode<HDrawable>
 	 * @return This drawable.
 	 */
 	public HDrawable rotateRad(float rad) {
-		_rotationRad += rad;
+		return rotationRad(rotationRad() + rad);
+	}
+	
+	public HDrawable rotatesChildren(boolean b) {
+		_flags = HMath.setBits(_flags, BITMASK_ROTATES_CHILDREN, b);
 		return this;
 	}
 	
-	// TODO rotatesChildren
+	public boolean rotatesChildren() {
+		return HMath.hasBits(_flags, BITMASK_ROTATES_CHILDREN);
+	}
 	
 	
 	// VISIBILITY //